第三次修正符号位扩展导致转义函数工作不正常的问题。

+++ qiniu/base.c
@@ -109,7 +109,7 @@ static char* Qiniu_escape(const char* s, escapeMode mode, Qiniu_Bool* fesc)
 
 	for (i = 0; i < len; i++) {
         // prevent c from sign extension
-		c = s[i] & 0xFF;
+		c = ((int)s[i]) & 0xFF;
 		if (Qiniu_shouldEscape(c, mode)) {
 			if (c == ' ' && mode == encodeQueryComponent) {
 				spaceCount++;
@@ -128,7 +128,7 @@ static char* Qiniu_escape(const char* s, escapeMode mode, Qiniu_Bool* fesc)
 	j = 0;
 	for (i = 0; i < len; i++) {
         // prevent c from sign extension
-		c = s[i] & 0xFF;
+		c = ((int)s[i]) & 0xFF;
 		if (Qiniu_shouldEscape(c, mode)) {
 			if (c == ' ' && mode == encodeQueryComponent) {
 				t[j] = '+';
